(refactor) move test from test-bronze folder to normal tests folder
