+++ .coveragerc
@@ -16,7 +16,7 @@ omit =
         hummingbot/connector/derivative/position.py
         hummingbot/connector/exchange/bitfinex*
         hummingbot/connector/exchange/foxbit*
-        hummingbot/connector/exchange/chainflip_lp*
+        hummingbot/connector/exchange/chainflip_lp/*
         hummingbot/connector/derivative/dydx_v4_perpetual/*
         hummingbot/connector/derivative/dydx_v4_perpetual/data_sources/*
         hummingbot/connector/exchange/injective_v2/account_delegation_script.py
