+++ Makefile
@@ -16,6 +16,7 @@ test:
  	--exclude-dir="test/hummingbot/connector/exchange/coinbase_advance_trade" \
  	--exclude-dir="test/hummingbot/connector/exchange/ndax" \
  	--exclude-dir="test/hummingbot/connector/exchange/foxbit" \
+ 	--exclude-dir="test/hummingbot/connector/exchange/chainflip_lp" \
  	--exclude-dir="test/hummingbot/connector/derivative/dydx_v4_perpetual" \
  	--exclude-dir="test/hummingbot/connector/gateway/clob_spot/data_sources/dexalot" \
  	--exclude-dir="test/hummingbot/strategy/amm_arb" \
