+++ test/hummingbot/connector/exchange/chainflip_lp/test_chainflip_lp_api_order_book_data_source.py
@@ -1,5 +1,6 @@
 import asyncio
 import re
+from test.hummingbot.connector.exchange.chainflip_lp.mock_rpc_executor import MockRPCExecutor
 from typing import Awaitable, Optional, Union
 from unittest import TestCase
 
@@ -14,8 +15,6 @@ from hummingbot.connector.exchange.chainflip_lp.chainflip_lp_api_order_book_data
 from hummingbot.connector.exchange.chainflip_lp.chainflip_lp_data_formatter import DataFormatter
 from hummingbot.connector.exchange.chainflip_lp.chainflip_lp_exchange import ChainflipLpExchange
 from hummingbot.connector.test_support.network_mocking_assistant import NetworkMockingAssistant
-
-from .mock_rpc_executor import MockRPCExecutor
 
 
 class TestChanflipLPAPIOrderBookDataSource(TestCase):
