+++ test/hummingbot/connector/exchange/chainflip_lp/test_chainflip_lp_data_source.py
@@ -1,4 +1,5 @@
 import asyncio
+from test.hummingbot.connector.exchange.chainflip_lp.mock_rpc_executor import MockRPCExecutor
 from typing import Awaitable
 from unittest import TestCase
 
@@ -9,8 +10,6 @@ from hummingbot.client.config.config_helpers import ClientConfigAdapter
 from hummingbot.connector.exchange.chainflip_lp.chainflip_lp_data_source import ChainflipLpDataSource
 from hummingbot.connector.exchange.chainflip_lp.chainflip_lp_exchange import ChainflipLpExchange
 from hummingbot.connector.test_support.network_mocking_assistant import NetworkMockingAssistant
-
-from .mock_rpc_executor import MockRPCExecutor
 
 
 class ChainflipLpDataSourceTests(TestCase):
